requirements.txt
Django==2.2
django-cors-headers
djangorestframework
djangorestframework_simplejwt
markdown
django-filter
pika